pyproject.toml
[tool.poetry]
name = "pythonx"
version = "0.1.0"
description = ""
authors = ["Dane Summers <dsummersl@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pynvim = "^0.4.3"
black = "^21.7b0"
python-vim = "^0.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
